feat: Neuron Driver 2.3.3.0
Source code extracted from file: aws-neuron-dkms-2.3.3.0.noarch.rpm
RPM Downloaded from repository: https://yum.repos.neuron.amazonaws.com


Release Notes
-------------
Date: 04/29/2022

New in this release
^^^^^^^^^^^^^^^^^^^

- Minor performance improvements on inference and loading of models.

Bug Fixes
^^^^^^^^^

- Reduced Host CPU usage when reading ``hw_counters`` metric from neuron-monitor
- Minor bug fixes. 
-------------


Metadata
--------
Package: aws-neuron-dkms
Version: 2.3.3.0
License: Unknown
Summary: aws-neuron 2.3.3.0 dkms package
Description: Kernel modules for aws-neuron 2.3.3.0 in a DKMS wrapper.
Filename: aws-neuron-dkms-2.3.3.0.noarch.rpm
Checksum: abfe5986a96cd7232d11ecb1e147a3a2a4dd5b362a09cf327d71c2fee03747c4
Buildhost: 34e086e6b437
Buildtime: 2022-04-13 18:14:41 +0000 UTC
GPG key primary uid: Amazon AWS Neuron <neuron-maintainers@amazon.com>
GPG key creation time: 2019-11-11 17:29:27 +0000 UTC
GPG key fingerprint: 00FA2C1079260870A76D2C285749CAD8646D9185
GPG check: OK
SHA256 check: OK
--------


Files
-----
- M: src/neuron_cdev.c
- M: src/neuron_core.h
- M: src/neuron_metrics.c
- M: src/neuron_trace.h
- M: src/v1/dma.c
- M: src/v2/dma.c
- M: src/neuron_fw_io.c
- M: src/neuron_mempool.h
- M: src/v2/notific.h
- M: src/neuron_device.h
- M: src/neuron_mempool.c
- M: src/neuron_metrics.h
- M: src/neuron_reg_access.h
- M: src/neuron_ring.h
- M: src/udma/udma_m2m.c
- M: src/udma/udma_main.c
- M: src/neuron_ioctl.h
- M: src/neuron_p2p.c
- M: src/neuron_p2p.h
- M: src/neuron_pid.c
- M: src/udma/udma.h
- M: src/v1/fw_io.c
- ?: archive/rpm/aws-neuron-dkms-2.3.3.0.noarch.rpm
- ?: archive/rpm/aws-neuron-dkms-2.3.3.0.noarch.rpm.sha256
- M: src/neuron_mmap.c
- M: src/share/neuron_driver_shared.h
- M: src/v2/notific.c
- M: src/neuron_topsp.c
- M: src/neuron_ds.c
- M: src/neuron_fw_io.h
- M: src/neuron_module.c
- M: src/neuron_pci.c
- M: src/dkms.conf
- M: src/neuron_dma.c
- M: src/neuron_ring.c
- M: src/v2/address_map.h
-----

+++ src/neuron_mempool.h
@@ -119,7 +119,7 @@ struct mem_chunk {
 	phys_addr_t pa; // physical address of the chunk
 	void *va; // virtual address of the chunk
 
-	u32 size; // chunk size
+	u64 size; // chunk size
 
 	struct mempool *mp; // backpointer to mp
 	struct mempool_set *mpset; // back pointer to mpset
@@ -182,7 +182,7 @@ struct mem_chunk *mpset_search_mc(struct mempool_set *mp, phys_addr_t pa);
  *
  * Return: 0 if allocation succeeds, a negative error code otherwise.
  */
-int mc_alloc(struct neuron_device *nd, enum mc_lifespan lifespan, u32 size,
+int mc_alloc(struct neuron_device *nd, enum mc_lifespan lifespan, u64 size,
 	     enum mem_location location, u32 channel, u32 region, u32 nc_id,
 	     struct mem_chunk **result);
 
@@ -200,7 +200,7 @@ int mc_alloc(struct neuron_device *nd, enum mc_lifespan lifespan, u32 size,
  *
  * Return: 0 if allocation succeeds, a negative error code otherwise.
  */
-int mc_alloc_align(struct neuron_device *nd, enum mc_lifespan lifespan, u32 size, u32 align,
+int mc_alloc_align(struct neuron_device *nd, enum mc_lifespan lifespan, u64 size, u64 align,
 		   enum mem_location location, u32 channel, u32 region, u32 nc_id,
 		   struct mem_chunk **result);
 /**
